Update main.yml
the ansible error is:
V2RAY_CFG_NAME is undefined. 
I remove this variable. 
I checked the v2ray.sh file  and it has a variable as configName that related to find command. Actually, it is not needed. Finally I defined it null.
Please check your v2ray.sh file.